+++ WeatherClientSwiftUI/WeatherClientSwiftUI/Weather/View/WeatherView.swift
@@ -20,7 +20,7 @@ struct WeatherView: View {
                 Text("Coordinates").tag(1)
             }
             .pickerStyle(.segmented)
-          
+            
             .onChange(of: segmentValue) { oldValue, newValue in
                 print("new value: \(newValue)")
             }
@@ -29,7 +29,7 @@ struct WeatherView: View {
                 TextField( "city name", text: $viewModel.city)
                     .padding(10)
                     .textFieldStyle(.roundedBorder)
-                    
+                
             } else {
                 TextField( "latitude", text: $viewModel.latitude)
                     .padding(10)
@@ -43,13 +43,18 @@ struct WeatherView: View {
                 if segmentValue == 0 {
                     print(viewModel.city)
                     Task {
-                        await viewModel.fetchWeather(byCyti: viewModel.city)
+                        await viewModel.networkService.fetchWeather(byCyti: viewModel.city) { result in
+                            viewModel.updateUI(with: result)
+                        }
                         
                     }
                 } else {
                     print("Lat: \(viewModel.latitude), lon: \(viewModel.longitude)")
                     Task {
-                        await viewModel.fetchWeather(byLatitude: viewModel.latitude, byLongitude: viewModel.longitude)
+                        await viewModel.networkService.fetchWeather(byLatitude: viewModel.latitude, byLongitude: viewModel.longitude) { result in
+                            viewModel.updateUI(with: result)
+                        }
+                        
                     }
                     
                 }
@@ -66,16 +71,10 @@ struct WeatherView: View {
             
             Spacer()
             
-            
-            
-            
-            
         }
         .padding()
         
     }
-    
-    
 }
 
 #Preview {
