+++ WeatherClientSwiftUI/WeatherClientSwiftUI/Weather/WeatherViewModel.swift
@@ -10,6 +10,7 @@ import Foundation
 @MainActor
 class WeatherViewModel: ObservableObject {
     
+    @Published var cityName: String = "Loading.."
     @Published var temperature: String = "--"
     @Published var pressure: String = "0.0"
     @Published var humidity: String = "0.0"
@@ -20,85 +21,88 @@ class WeatherViewModel: ObservableObject {
     @Published var latitude: String = ""
     @Published var longitude: String = ""
     
- 
-    func fetchWeather(byCyti city: String) async {
-        let linkApi = ConstantLink.mainLink.rawValue + ConstantLink.city.rawValue + "\(city)" + "&" + ConstantLink.appid.rawValue + "&" + ConstantLink.metrics.rawValue
-       
-        
-        guard let url = URL(string: linkApi) else {
-            assertionFailure("Wrong link Api \(linkApi)")
-            return
-        }
-        
-        var requestUrl = URLRequest(url: url)
-        requestUrl.httpMethod = "GET"
-        
-        let task = URLSession.shared.dataTask(with: requestUrl) { data, response, error in
-            if let responseError = error {
-                assertionFailure("\(responseError)")
-            }
-            
-            guard let responseData = data else {
-                assertionFailure("Problem with data")
-                return
-            }
-            
-            do {
-                let decodeData = try JSONDecoder().decode(WeatherResult.self, from: responseData)
-                
-                DispatchQueue.main.async {
-                    self.updateUI(with: decodeData)
-                }
-                
-            } catch(let parseError){
-                print(parseError)
-            }
-        }
-        task.resume()
-  
-        print("Fetching weather for city: \(city)")
-    }
+    let networkService = NetworkService.shared
     
-    func fetchWeather(byLatitude latitude: String, byLongitude longitude: String) async {
-        
-        let linkApi = ConstantLink.mainLink.rawValue + ConstantLink.latitude.rawValue + "\(latitude)" + "&" + ConstantLink.longitude.rawValue + "\(longitude)" + "&" + ConstantLink.appid.rawValue + "&" + ConstantLink.metrics.rawValue
-        
-        guard let url = URL(string: linkApi) else {
-            assertionFailure("Wring url link: \(linkApi)")
-            return
-        }
-        
-        var requestUrl = URLRequest(url: url)
-        requestUrl.httpMethod = "GET"
-        
-        let task = URLSession.shared.dataTask(with: requestUrl) { data, response, error in
-            if let responseError = error {
-                assertionFailure("\(responseError)")
-            }
-            
-            guard let responseData = data else {
-                assertionFailure("No data")
-                return
-            }
-            
-            do {
-                let decodeData = try JSONDecoder().decode(WeatherResult.self, from: responseData)
-                print(decodeData)
-                
-                DispatchQueue.main.async {
-                    self.updateUI(with: decodeData)
-                }
-                
-            } catch(let parseError) {
-                print(parseError)
-            }
-        }
-        task.resume()
-        
-        print("Fetching weather for latitude: \(latitude), longitude: \(longitude)")
-    }
+ 
+//    func fetchWeather(byCyti city: String) async {
+//        let linkApi = ConstantLink.mainLink.rawValue + ConstantLink.city.rawValue + "\(city)" + "&" + ConstantLink.appid.rawValue + "&" + ConstantLink.metrics.rawValue
+//       
+//        
+//        guard let url = URL(string: linkApi) else {
+//            assertionFailure("Wrong link Api \(linkApi)")
+//            return
+//        }
+//        
+//        var requestUrl = URLRequest(url: url)
+//        requestUrl.httpMethod = "GET"
+//        
+//        let task = URLSession.shared.dataTask(with: requestUrl) { data, response, error in
+//            if let responseError = error {
+//                assertionFailure("\(responseError)")
+//            }
+//            
+//            guard let responseData = data else {
+//                assertionFailure("Problem with data")
+//                return
+//            }
+//            
+//            do {
+//                let decodeData = try JSONDecoder().decode(WeatherResult.self, from: responseData)
+//                
+//                DispatchQueue.main.async {
+//                    self.updateUI(with: decodeData)
+//                }
+//                
+//            } catch(let parseError){
+//                print(parseError)
+//            }
+//        }
+//        task.resume()
+//  
+//        print("Fetching weather for city: \(city)")
+//    }
+//    
+//    func fetchWeather(byLatitude latitude: String, byLongitude longitude: String) async {
+//        
+//        let linkApi = ConstantLink.mainLink.rawValue + ConstantLink.latitude.rawValue + "\(latitude)" + "&" + ConstantLink.longitude.rawValue + "\(longitude)" + "&" + ConstantLink.appid.rawValue + "&" + ConstantLink.metrics.rawValue
+//        
+//        guard let url = URL(string: linkApi) else {
+//            assertionFailure("Wring url link: \(linkApi)")
+//            return
+//        }
+//        
+//        var requestUrl = URLRequest(url: url)
+//        requestUrl.httpMethod = "GET"
+//        
+//        let task = URLSession.shared.dataTask(with: requestUrl) { data, response, error in
+//            if let responseError = error {
+//                assertionFailure("\(responseError)")
+//            }
+//            
+//            guard let responseData = data else {
+//                assertionFailure("No data")
+//                return
+//            }
+//            
+//            do {
+//                let decodeData = try JSONDecoder().decode(WeatherResult.self, from: responseData)
+//                print(decodeData)
+//                
+//                DispatchQueue.main.async {
+//                    self.updateUI(with: decodeData)
+//                }
+//                
+//            } catch(let parseError) {
+//                print(parseError)
+//            }
+//        }
+//        task.resume()
+//        
+//        print("Fetching weather for latitude: \(latitude), longitude: \(longitude)")
+//    }
 
-    private func updateUI(with weather: WeatherResult) {
+    func updateUI(with weather: WeatherResult) {
+        cityName = "\(weather.name)"
         temperature = "\(weather.main.temp) C"
         pressure = "\(weather.main.pressure)"
         humidity = "\(weather.main.humidity)"
